+++ src/main.cpp
@@ -72,5 +72,5 @@ $on_mod (Loaded){
 }
 
 $on_game(Exiting) {
-	Data::flushCache(); // save data on shutdown
+	//Data::flushCache(); // save data on shutdown (crashes game, fix that sometime idk)
 }