+++ changelog.md
@@ -1,7 +1,8 @@
 # v1.2.0
-- made it so leveldata is now in a seperate file for each level instead of one big file to prevent lag when saving / loading
+- made it so level data is now in a separate file for each level instead of one big file to prevent lag when saving / loading
 - added level cache to reduce unnecessary read operations when opening the menu popup
 - made loading the data asynchronous to prevent the game freezing while loading the menu popup
+- made the pause menu popup update when not using 'Remove pauses from playtime'
 - fixed an issue where 'remove pauses when AFK' would not remove pauses when exiting or resuming in certain ways
 # v1.1.0
 - Ported to 2.2081